use BasePluginXmlTest
