+++ plugins/com.google.cloud.tools.eclipse.test.util/src/com/google/cloud/tools/eclipse/test/util/BasePluginXmlTest.java
@@ -16,13 +16,20 @@
 
 package com.google.cloud.tools.eclipse.test.util;
 
+import org.eclipse.core.runtime.IExtensionPoint;
+import org.eclipse.core.runtime.IExtensionRegistry;
+import org.eclipse.core.runtime.RegistryFactory;
+import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Rule;
+import org.junit.Test;
 import org.w3c.dom.Document;
+import org.w3c.dom.Element;
+import org.w3c.dom.NodeList;
 
 public class BasePluginXmlTest {
 
-  @Rule public PluginXmlDocument pluginXmlDocument = new PluginXmlDocument();
+  @Rule public final PluginXmlDocument pluginXmlDocument = new PluginXmlDocument();
   private Document doc;
 
   @Before
@@ -30,7 +37,30 @@ public class BasePluginXmlTest {
     doc = pluginXmlDocument.get();
   }
 
-  protected Document getDoc() {
+  protected Document getDocument() {
     return doc;
   }
+  
+  // Generic tests that should be true of all plugin.xml files
+  
+  @Test
+  public void testRootElementIsPlugin() {
+    Assert.assertEquals("plugin", getDocument().getDocumentElement().getNodeName());
+  }
+   
+  @Test
+  public void testLoadExtensionPoints() {
+    NodeList extensions = getDocument().getDocumentElement().getElementsByTagName("extension");
+    
+    // todo should we test that the file has at least one extension point?
+    
+    for (int i = 0; i < extensions.getLength(); i++) {
+      Element extension = (Element) extensions.item(i);
+      String point = extension.getAttribute("point");
+      IExtensionRegistry registry = RegistryFactory.getRegistry();
+      IExtensionPoint extensionPoint = registry.getExtensionPoint(point);
+      Assert.assertNotNull(extensionPoint);
+    }
+  }
+  
 }
